bugfix: 修复一些bug及更正一些错误的配置文件

+++ intellidesk-user-service/src/main/java/com/intellidesk/user/domain/dto/DepartmentQueryRequest.java
@@ -32,10 +32,10 @@ public class DepartmentQueryRequest implements Serializable {
     /**
      * 页码
      */
-    private Integer pageNum = 1;
+    private Long pageNum = 1L;
 
     /**
      * 每页大小
      */
-    private Integer pageSize = 10;
+    private Long pageSize = 10L;
 }
